SOLR-15488: Move v2 API impl's into uniform subpackages
As support for POJO/annotation-based v2 APIs spread, these classes will
become more and more numerous, polluting other packages.  This commit
creates a series of 'api' subpackages in an effort to keep the API code
physically separate.
