+++ solr/core/src/java/org/apache/solr/packagemanager/PackageUtils.java
@@ -43,7 +43,7 @@ import org.apache.solr.common.params.ModifiableSolrParams;
 import org.apache.solr.common.util.Utils;
 import org.apache.solr.core.BlobRepository;
 import org.apache.solr.filestore.DistribPackageStore;
-import org.apache.solr.filestore.PackageStoreAPI;
+import org.apache.solr.filestore.api.PackageStoreAPI;
 import org.apache.solr.packagemanager.SolrPackage.Manifest;
 import org.apache.solr.util.SolrJacksonAnnotationInspector;
