+++ solr/core/src/java/org/apache/solr/pkg/api/PackageAPI.java
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.solr.pkg;
+package org.apache.solr.pkg.api;
 
 import java.io.IOException;
 import java.io.StringWriter;
@@ -41,7 +41,8 @@ import org.apache.solr.common.util.CommandOperation;
 import org.apache.solr.common.util.ReflectMapWriter;
 import org.apache.solr.common.util.Utils;
 import org.apache.solr.core.CoreContainer;
-import org.apache.solr.filestore.PackageStoreAPI;
+import org.apache.solr.filestore.api.PackageStoreAPI;
+import org.apache.solr.pkg.PackageLoader;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
 import org.apache.solr.util.SolrJacksonAnnotationInspector;
@@ -94,6 +95,8 @@ public class PackageAPI {
       SolrZkClient.checkInterrupted(e);
     }
   }
+
+  public Packages getPackages() { return pkgs; }
 
   private void registerListener(SolrZkClient zkClient)
       throws KeeperException, InterruptedException {
