+++ solr/core/src/test/org/apache/solr/handler/admin/api/V2CollectionsAPIMappingTest.java
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.handler.admin;
+package org.apache.solr.handler.admin.api;
 
 import com.google.common.collect.Maps;
 import org.apache.solr.SolrTestCaseJ4;
@@ -35,7 +35,9 @@ import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.common.util.CommandOperation;
 import org.apache.solr.common.util.ContentStreamBase;
 import org.apache.solr.core.backup.BackupManager;
-import org.apache.solr.handler.CollectionsAPI;
+import org.apache.solr.handler.admin.CollectionsHandler;
+import org.apache.solr.handler.admin.api.CollectionsAPI;
+import org.apache.solr.handler.admin.api.TestCollectionAPIs;
 import org.apache.solr.request.LocalSolrQueryRequest;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
@@ -56,7 +58,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
 /**
- * Unit tests for the API mappings found in {@link org.apache.solr.handler.CollectionsAPI}.
+ * Unit tests for the API mappings found in {@link CollectionsAPI}.
  *
  * This test bears many similarities to {@link TestCollectionAPIs} which appears to test the mappings indirectly by
  * checking message sent to the ZK overseer (which is similar, but not identical to the v1 param list).  If there's no
